Put sql api behind /sql
